# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MyHR API",
...
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {
      "masterApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Master API key issued to 1tap (env `MASTER_API_KEY`). Sent as `Authorization: Bearer mh_live_…`. Master callers may operate across all tenants and must send `X-Tenant-Id` to scope tenant-specific calls."
      },
      "tenantApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Org-scoped API key minted from the dashboard. Sent as `Authorization: Bearer mh_live_…`. The org id is derived from the key, so `X-Tenant-Id` is ignored."
      },
      "userSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Better Auth session token forwarded by the web app. End-user callers send `X-Org-Id` to select which of their orgs the request targets. `X-Actor` is ignored on user-session requests (the actor is taken from the session to prevent spoofing)."
      }
    },
...
  },
  "paths": {
...
    "/v1/employees/{id}/export": {
      "get": {
        "operationId": "exportEmployee",
        "summary": "Export employee data (GDPR Art. 15)",
        "tags": [
          "Employees"
        ],
        "description": "Returns the full employee record. A follow-up will also bundle contracts and documents into a zip.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "example": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "description": "Tenant org id. Required on all tenant-scoped endpoints."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-actor",
            "required": false,
            "description": "Optional JSON `{ id?, email?, name? }` attributing this request to a specific 1tap user for the audit log."
          }
        ],
        "security": [
          {
            "masterApiKey": []
          },
          {
            "tenantApiKey": []
          },
          {
            "userSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "employee": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "orgId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "externalId": {
                          "type": "string",
                          "nullable": true
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "preferredName": {
                          "type": "string",
                          "nullable": true
                        },
                        "jobTitle": {
                          "type": "string",
                          "nullable": true
                        },
                        "department": {
                          "type": "string",
                          "nullable": true
                        },
                        "managerId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "country": {
                          "type": "string",
                          "enum": [
                            "us",
                            "de"
                          ]
                        },
                        "startDate": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                        },
                        "endDate": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "onboarding",
                            "active",
                            "on_leave",
                            "terminated"
                          ]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "orgId",
                        "externalId",
                        "email",
                        "firstName",
                        "lastName",
                        "preferredName",
                        "jobTitle",
                        "department",
                        "managerId",
                        "country",
                        "startDate",
                        "endDate",
                        "status",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "exportedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "employee",
                    "exportedAt"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "employee": {
                    "id": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
                    "orgId": "11111111-2222-3333-4444-555555555555",
                    "externalId": "emp_4271",
                    "email": "ada@acme.com",
                    "firstName": "Ada",
                    "lastName": "Lovelace",
                    "preferredName": null,
                    "jobTitle": "Staff Engineer",
                    "department": "Engineering",
                    "managerId": null,
                    "country": "us",
                    "startDate": "2026-06-01",
                    "endDate": null,
                    "status": "onboarding",
                    "createdAt": "2026-05-04T12:00:00.000Z",
                    "updatedAt": "2026-05-04T12:00:00.000Z"
                  },
                  "exportedAt": "2026-05-04T12:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request shape error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Validation or request shape error."
                },
                "example": {
                  "error": {
                    "code": "bad_request",
                    "message": "managerId does not reference an employee in this tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Missing or invalid API key."
                },
                "example": {
                  "error": {
                    "code": "unauthorized",
                    "message": "Missing or invalid API key"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Caller is authenticated but not allowed to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Caller is authenticated but not allowed to perform this action."
                },
                "example": {
                  "error": {
                    "code": "forbidden",
                    "message": "Caller is not allowed to perform this action"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Resource not found."
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Not found"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Retry after the seconds indicated by the `Retry-After` header.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Rate limit exceeded. Retry after the seconds indicated by the `Retry-After` header."
                },
                "example": {
                  "error": {
                    "code": "rate_limited",
                    "message": "Rate limit exceeded. Retry after 6s.",
                    "details": {
                      "retryAfterSec": 6
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Machine-readable error code, e.g. `not_found`."
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Unexpected server error."
                },
                "example": {
                  "error": {
                    "code": "internal_error",
                    "message": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "bash",
            "label": "curl",
            "source": "curl -X GET 'https://api.myhr.example/v1/employees/aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee/export' \\\n  -H 'Authorization: Bearer $MYHR_API_KEY' \\\n  -H 'X-Tenant-Id: 11111111-2222-3333-4444-555555555555'"
          },
          {
            "lang": "javascript",
            "label": "@myhr/sdk",
            "source": "import { createClient } from \"@myhr/sdk\";\n\nconst myhr = createClient({\n  baseUrl: \"https://api.myhr.example\",\n  getToken: () => process.env.MYHR_API_KEY!,\n  defaultTenantId: \"11111111-2222-3333-4444-555555555555\",\n});\n\nawait myhr.employees.exportData(\"aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee\");"
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.myhr.example",
      "description": "test"
    }
  ],
  "security": [
    {
      "masterApiKey": []
    },
    {
      "tenantApiKey": []
    },
    {
      "userSession": []
    }
  ],
...
}